[241021] refactor: - myapis-list 리팩토링 (컴포넌트 분리 및 코드 분리)
